Pulls Latlong from Petfinder API
This version works, but still displays the nulls.  I tried to merge the changes I made with the newest file and it broke my code for reasons I could not discern.  If there was a way to keep my updates to this older version petfinder-api.js file while adding in whatever code Ben wrote to ignore the 'null' addresses we'd be home free.